Add feature engine for indicators
